Final project

+++ MathLib/src/ru/vsu/cs/zhilyaev/MathForVectors2.java
@@ -9,11 +9,11 @@ public class MathForVectors2 {
         return new Vector2(v1.getX() - v2.getX(), v1.getY() - v2.getY());
     }
 
-    public Vector2 increaseVectorOnNumber(Vector2 v1, double n) {
+    public Vector2 increaseVectorOnNumber(Vector2 v1, float n) {
         return new Vector2(v1.getX() * n, v1.getY() * n);
     }
 
-    public Vector2 divisionVectorOnNumber(Vector2 v1, double n) {
+    public Vector2 divisionVectorOnNumber(Vector2 v1, float n) {
         if(n == 0) {
             System.out.println("На 0 делить нельзя");
             return null;
@@ -22,16 +22,16 @@ public class MathForVectors2 {
         return new Vector2(v1.getX() / n, v1.getY() / n);
     }
 
-    public double lengthOfVector(Vector2 v1) {
-        return Math.sqrt(v1.getX() * v1.getX() + v1.getY() * v1.getY());
+    public float lengthOfVector(Vector2 v1) {
+        return (float) Math.sqrt(v1.getX() * v1.getX() + v1.getY() * v1.getY());
     }
 
-    public double scalarProduct(Vector2 v1, Vector2 v2) {
+    public float scalarProduct(Vector2 v1, Vector2 v2) {
         return v1.getX() * v2.getX() + v1.getY() * v2.getY();
     }
 
     public Vector2 normalization(Vector2 v1) {
-        double invLength = 1 / (Math.sqrt(v1.getX() * v1.getX() + v1.getY() * v1.getY()));
+        float invLength = (float) (1 / (Math.sqrt(v1.getX() * v1.getX() + v1.getY() * v1.getY())));
         return new Vector2(v1.getX() * invLength, v1.getY() * invLength);
     }
 }
