+++ MathLib/src/ru/vsu/cs/zhilyaev/MathForVectors2Test.java
@@ -7,11 +7,11 @@ class MathForVectors2Test {
 
     @Test
     void sumOfVectors() {
-        Vector2 v1 = new Vector2(1.0, 2.0);
-        Vector2 v2 = new Vector2(5.0, 6.0);
+        Vector2 v1 = new Vector2(1, 2);
+        Vector2 v2 = new Vector2(5, 6);
 
         Vector2 vectorFromMethod = mathForVectors2.sumOfVectors(v1, v2);
-        Vector2 vectorResult = new Vector2(6.0, 8.0);
+        Vector2 vectorResult = new Vector2(6, 8);
 
         double eps = 1e-6;
 
@@ -21,11 +21,11 @@ class MathForVectors2Test {
 
     @Test
     void diffOfVectors() {
-        Vector2 v1 = new Vector2(1.0, 2.0);
-        Vector2 v2 = new Vector2(5.0, 6.0);
+        Vector2 v1 = new Vector2(1, 2);
+        Vector2 v2 = new Vector2(5, 6);
 
         Vector2 vectorFromMethod = mathForVectors2.diffOfVectors(v1, v2);
-        Vector2 vectorResult = new Vector2(-4.0, -4.0);
+        Vector2 vectorResult = new Vector2(-4, -4);
 
         double eps = 1e-6;
 
@@ -35,11 +35,11 @@ class MathForVectors2Test {
 
     @Test
     void increaseVectorOnNumber() {
-        Vector2 v1 = new Vector2(1.0, 2.0);
-        double n = 5.0;
+        Vector2 v1 = new Vector2(1, 2);
+        float n = 5.0f;
 
         Vector2 vectorFromMethod = mathForVectors2.increaseVectorOnNumber(v1, n);
-        Vector2 vectorResult = new Vector2(5.0, 10.0);
+        Vector2 vectorResult = new Vector2(5, 10);
 
         double eps = 1e-6;
 
@@ -49,8 +49,8 @@ class MathForVectors2Test {
 
     @Test
     void divisionVectorOnZero() {
-        Vector2 v1 = new Vector2(1.0, 2.0);
-        double n = 0.0;
+        Vector2 v1 = new Vector2(1, 2);
+        float n = 0;
 
         Vector2 vectorFromMethod = mathForVectors2.divisionVectorOnNumber(v1, n);
         assert vectorFromMethod == null;
@@ -58,11 +58,11 @@ class MathForVectors2Test {
 
     @Test
     void divisionVectorOnNumber() {
-        Vector2 v1 = new Vector2(6.0, 8.0);
-        double n = 2;
+        Vector2 v1 = new Vector2(6, 8);
+        float n = 2;
 
         Vector2 vectorFromMethod = mathForVectors2.divisionVectorOnNumber(v1, n);
-        Vector2 vectorResult = new Vector2(3.0, 4.0);
+        Vector2 vectorResult = new Vector2(3, 4);
 
         double eps = 1e-6;
 
@@ -72,10 +72,10 @@ class MathForVectors2Test {
 
     @Test
     void lengthOfVector() {
-        Vector2 v1 = new Vector2(6.0, 8.0);
+        Vector2 v1 = new Vector2(6, 8);
 
-        double lengthFromMethod = mathForVectors2.lengthOfVector(v1);
-        double lengthResult = 10.0;
+        float lengthFromMethod = mathForVectors2.lengthOfVector(v1);
+        float lengthResult = 10.0f;
 
         double eps = 1e-6;
 
@@ -84,8 +84,8 @@ class MathForVectors2Test {
 
     @Test
     void scalarProduct() {
-        Vector2 v1 = new Vector2(6.0, 8.0);
-        Vector2 v2 = new Vector2(5.0, 4.0);
+        Vector2 v1 = new Vector2(6, 8);
+        Vector2 v2 = new Vector2(5, 4);
 
         double resultFromMethod = mathForVectors2.scalarProduct(v1, v2);
         double result = 62;
@@ -97,10 +97,10 @@ class MathForVectors2Test {
 
     @Test
     void normalization() {
-        Vector2 v1 = new Vector2(6.0, 8.0);
+        Vector2 v1 = new Vector2(6, 8);
 
         Vector2 vectorFromMethod = mathForVectors2.normalization(v1);
-        Vector2 vectorResult = new Vector2(6.0 * 0.1,8.0 * 0.1);
+        Vector2 vectorResult = new Vector2(6 * 0.1f,8 * 0.1f);
 
         double eps = 1e-6;
