+++ MathLib/src/ru/vsu/cs/zhilyaev/MathForVectors3.java
@@ -9,11 +9,11 @@ public class MathForVectors3 {
         return new Vector3(v1.getX() - v2.getX(), v1.getY() - v2.getY(), v1.getZ() - v2.getZ());
     }
 
-    public Vector3 increaseVectorOnNumber(Vector3 v1, double n) {
+    public Vector3 increaseVectorOnNumber(Vector3 v1, float n) {
         return new Vector3(v1.getX() * n, v1.getY() * n, v1.getZ() * n);
     }
 
-    public Vector3 divisionVectorOnNumber(Vector3 v1, double n) {
+    public Vector3 divisionVectorOnNumber(Vector3 v1, float n) {
         if(n == 0) {
             System.out.println("На 0 делить нельзя");
             return null;
@@ -22,23 +22,23 @@ public class MathForVectors3 {
         return new Vector3(v1.getX() / n, v1.getY() / n, v1.getZ() / n);
     }
 
-    public double lengthOfVector(Vector3 v1) {
-        return Math.sqrt(v1.getX() * v1.getX() + v1.getY() * v1.getY() + v1.getZ() * v1.getZ());
+    public float lengthOfVector(Vector3 v1) {
+        return (float) Math.sqrt(v1.getX() * v1.getX() + v1.getY() * v1.getY() + v1.getZ() * v1.getZ());
     }
 
-    public double scalarProduct(Vector3 v1, Vector3 v2) {
+    public float scalarProduct(Vector3 v1, Vector3 v2) {
         return v1.getX() * v2.getX() + v1.getY() * v2.getY() + v1.getZ() * v2.getZ();
     }
 
     public Vector3 normalization(Vector3 v1) {
-        double invLength = 1 / Math.sqrt(v1.getX() * v1.getX() + v1.getY() * v1.getY() + v1.getZ() * v1.getZ());
+        float invLength = (float) (1 / Math.sqrt(v1.getX() * v1.getX() + v1.getY() * v1.getY() + v1.getZ() * v1.getZ()));
         return new Vector3(v1.getX() * invLength, v1.getY() * invLength, v1.getZ() * invLength);
     }
 
     public Vector3 vectorProduct(Vector3 v1, Vector3 v2) {
-        double x = v1.getY() * v2.getZ() - v1.getZ() * v2.getY();
-        double y = -1 * (v1.getX() * v2.getZ() - v1.getZ() * v2.getX());
-        double z = v1.getX() * v2.getY() - v1.getY() * v2.getX();
+        float x = v1.getY() * v2.getZ() - v1.getZ() * v2.getY();
+        float y = -1 * (v1.getX() * v2.getZ() - v1.getZ() * v2.getX());
+        float z = v1.getX() * v2.getY() - v1.getY() * v2.getX();
 
         return new Vector3(x, y ,z);
     }
