+++ MathLib/src/ru/vsu/cs/zhilyaev/MathForVectors3Test.java
@@ -7,11 +7,11 @@ class MathForVectors3Test {
 
     @Test
     void sumOfVectors() {
-        Vector3 v1 = new Vector3(1.0, 2.0, 7.0);
-        Vector3 v2 = new Vector3(5.0, 6.0, 3.0);
+        Vector3 v1 = new Vector3(1.0f, 2.0f, 7.0f);
+        Vector3 v2 = new Vector3(5.0f, 6.0f, 3.0f);
 
         Vector3 vectorFromMethod = mathForVectors3.sumOfVectors(v1, v2);
-        Vector3 vectorResult = new Vector3(6.0, 8.0, 10.0);
+        Vector3 vectorResult = new Vector3(6.0f, 8.0f, 10.0f);
 
         double eps = 1e-6;
 
@@ -22,11 +22,11 @@ class MathForVectors3Test {
 
     @Test
     void diffOfVectors() {
-        Vector3 v1 = new Vector3(1.0, 2.0, 8.0);
-        Vector3 v2 = new Vector3(5.0, 6.0, 3.5);
+        Vector3 v1 = new Vector3(1.0f, 2.0f, 8.0f);
+        Vector3 v2 = new Vector3(5.0f, 6.0f, 3.5f);
 
         Vector3 vectorFromMethod = mathForVectors3.diffOfVectors(v1, v2);
-        Vector3 vectorResult = new Vector3(-4.0, -4.0, 4.5);
+        Vector3 vectorResult = new Vector3(-4.0f, -4.0f, 4.5f);
 
         double eps = 1e-6;
 
@@ -37,11 +37,11 @@ class MathForVectors3Test {
 
     @Test
     void increaseVectorOnNumber() {
-        Vector3 v1 = new Vector3(1.0, 2.0, 5.2);
-        double n = 5.0;
+        Vector3 v1 = new Vector3(1.0f, 2.0f, 5.2f);
+        float n = 5.0f;
 
         Vector3 vectorFromMethod = mathForVectors3.increaseVectorOnNumber(v1, n);
-        Vector3 vectorResult = new Vector3(5.0, 10.0, 26.0);
+        Vector3 vectorResult = new Vector3(5.0f, 10.0f, 26.0f);
 
         double eps = 1e-6;
 
@@ -52,8 +52,8 @@ class MathForVectors3Test {
 
     @Test
     void divisionVectorOnZero() {
-        Vector3 v1 = new Vector3(1.0, 2.0, 5.0);
-        double n = 0.0;
+        Vector3 v1 = new Vector3(1.0f, 2.0f, 5.0f);
+        float n = 0.0f;
 
         Vector3 vectorFromMethod = mathForVectors3.divisionVectorOnNumber(v1, n);
         assert vectorFromMethod == null;
@@ -61,11 +61,11 @@ class MathForVectors3Test {
 
     @Test
     void divisionVectorOnNumber() {
-        Vector3 v1 = new Vector3(6.0, 8.0, 12.2);
-        double n = 2;
+        Vector3 v1 = new Vector3(6.0f, 8.0f, 12.2f);
+        float n = 2;
 
         Vector3 vectorFromMethod = mathForVectors3.divisionVectorOnNumber(v1, n);
-        Vector3 vectorResult = new Vector3(3.0, 4.0, 6.1);
+        Vector3 vectorResult = new Vector3(3.0f, 4.0f, 6.1f);
 
         double eps = 1e-6;
 
@@ -76,10 +76,10 @@ class MathForVectors3Test {
 
     @Test
     void lengthOfVector() {
-        Vector3 v1 = new Vector3(2.0, 2.0, 1.0);
+        Vector3 v1 = new Vector3(2.0f, 2.0f, 1.0f);
 
-        double lengthFromMethod = mathForVectors3.lengthOfVector(v1);
-        double lengthResult = 3.0;
+        float lengthFromMethod = mathForVectors3.lengthOfVector(v1);
+        float lengthResult = 3.0f;
 
         double eps = 1e-6;
 
@@ -88,11 +88,11 @@ class MathForVectors3Test {
 
     @Test
     void scalarProduct() {
-        Vector3 v1 = new Vector3(6.0, 8.0, 2.0);
-        Vector3 v2 = new Vector3(5.0, 4.0, 7.5);
+        Vector3 v1 = new Vector3(6.0f, 8.0f, 2.0f);
+        Vector3 v2 = new Vector3(5.0f, 4.0f, 7.5f);
 
-        double resultFromMethod = mathForVectors3.scalarProduct(v1, v2);
-        double result = 77;
+        float resultFromMethod = mathForVectors3.scalarProduct(v1, v2);
+        float result = 77;
 
         double eps = 1e-6;
 
@@ -101,10 +101,10 @@ class MathForVectors3Test {
 
     @Test
     void normalization() {
-        Vector3 v1 = new Vector3(1.0, 8.0, 4.0);
+        Vector3 v1 = new Vector3(1.0f, 8.0f, 4.0f);
 
         Vector3 vectorFromMethod = mathForVectors3.normalization(v1);
-        Vector3 vectorResult = new Vector3(1.0 / 9.0,8.0 / 9.0, 4.0 / 9.0);
+        Vector3 vectorResult = new Vector3(1.0f / 9.0f,8.0f / 9.0f, 4.0f / 9.0f);
 
         double eps = 1e-6;
 
@@ -116,11 +116,11 @@ class MathForVectors3Test {
 
     @Test
     void vectorProduct() {
-        Vector3 v1 = new Vector3(6.0, 8.0, 2.0);
-        Vector3 v2 = new Vector3(5.0, 4.0, 7.5);
+        Vector3 v1 = new Vector3(6.0f, 8.0f, 2.0f);
+        Vector3 v2 = new Vector3(5.0f, 4.0f, 7.5f);
 
         Vector3 vectorFromMethod = mathForVectors3.vectorProduct(v1, v2);
-        Vector3 vectorResult = new Vector3(52.0, -35.0, -16.0);
+        Vector3 vectorResult = new Vector3(52.0f, -35.0f, -16.0f);
 
         double eps = 1e-6;
