+++ MathLib/src/ru/vsu/cs/zhilyaev/MathForVectors4Test.java
@@ -7,11 +7,11 @@ class MathForVectors4Test {
 
     @Test
     void sumOfVectors() {
-        Vector4 v1 = new Vector4(6.0, 5.2, 3.3, 1.0);
-        Vector4 v2 = new Vector4(2.2, 5.25, 2.0, 1.1);
+        Vector4 v1 = new Vector4(6.0f, 5.2f, 3.3f, 1.0f);
+        Vector4 v2 = new Vector4(2.2f, 5.25f, 2.0f, 1.1f);
 
         Vector4 vectorFromMethod = mathForVectors4.sumOfVectors(v1, v2);
-        Vector4 vectorResult = new Vector4(8.2, 10.45, 5.3, 2.1);
+        Vector4 vectorResult = new Vector4(8.2f, 10.45f, 5.3f, 2.1f);
 
         double eps = 1e-6;
 
@@ -23,11 +23,11 @@ class MathForVectors4Test {
 
     @Test
     void diffOfVectors() {
-        Vector4 v1 = new Vector4(6.0, 5.2, 3.3, 1.0);
-        Vector4 v2 = new Vector4(2.2, 5.25, 2.0, 1.1);
+        Vector4 v1 = new Vector4(6.0f, 5.2f, 3.3f, 1.0f);
+        Vector4 v2 = new Vector4(2.2f, 5.25f, 2.0f, 1.1f);
 
         Vector4 vectorFromMethod = mathForVectors4.diffOfVectors(v1, v2);
-        Vector4 vectorResult = new Vector4(6.0 - 2.2, 5.2 - 5.25, 3.3 - 2.0, 1.0 - 1.1);
+        Vector4 vectorResult = new Vector4(6.0f - 2.2f, 5.2f - 5.25f, 3.3f - 2.0f, 1.0f - 1.1f);
 
         double eps = 1e-6;
 
@@ -39,11 +39,11 @@ class MathForVectors4Test {
 
     @Test
     void increaseVectorOnNumber() {
-        Vector4 v1 = new Vector4(6.0, 5.2, 3.4, 1.1);
-        double n = 3.0;
+        Vector4 v1 = new Vector4(6.0f, 5.2f, 3.4f, 1.1f);
+        float n = 3.0f;
 
         Vector4 vectorFromMethod = mathForVectors4.increaseVectorOnNumber(v1, n);
-        Vector4 vectorResult = new Vector4(6.0 * 3.0, 5.2 * 3.0, 3.4 * 3.0, 1.1 * 3.0);
+        Vector4 vectorResult = new Vector4(6.0f * 3.0f, 5.2f * 3.0f, 3.4f * 3.0f, 1.1f * 3.0f);
 
         double eps = 1e-6;
 
@@ -55,11 +55,11 @@ class MathForVectors4Test {
 
     @Test
     void divisionOnNumber() {
-        Vector4 v1 = new Vector4(6.0, 5.2, 3.4, 1.0);
-        double n = 2.0;
+        Vector4 v1 = new Vector4(6.0f, 5.2f, 3.4f, 1.0f);
+        float n = 2.0f;
 
         Vector4 vectorFromMethod = mathForVectors4.divisionVectorOnNumber(v1, n);
-        Vector4 vectorResult = new Vector4(3.0, 2.6, 1.7, 0.5);
+        Vector4 vectorResult = new Vector4(3.0f, 2.6f, 1.7f, 0.5f);
 
         double eps = 1e-6;
 
@@ -71,8 +71,8 @@ class MathForVectors4Test {
 
     @Test
     void divisionOnZero() {
-        Vector4 v1 = new Vector4(6.0, 5.2, 3.4, 1.0);
-        double n = 0.0;
+        Vector4 v1 = new Vector4(6.0f, 5.2f, 3.4f, 1.0f);
+        float n = 0.0f;
 
         Vector4 vectorFromMethod = mathForVectors4.divisionVectorOnNumber(v1, n);
         assert vectorFromMethod == null;
@@ -80,10 +80,10 @@ class MathForVectors4Test {
 
     @Test
     void lengthOfVector() {
-        Vector4 v1 = new Vector4(1.0, 1.0, 5.0, 3.0);
+        Vector4 v1 = new Vector4(1.0f, 1.0f, 5.0f, 3.0f);
 
-        double lengthFromMethod = mathForVectors4.lengthOfVector(v1);
-        double lengthResult = 6.0;
+        float lengthFromMethod = mathForVectors4.lengthOfVector(v1);
+        float lengthResult = 6.0f;
 
         double eps = 1e-6;
 
@@ -92,11 +92,11 @@ class MathForVectors4Test {
 
     @Test
     void scalarProduct() {
-        Vector4 v1 = new Vector4(6.0, 5.2, 3.3, 1.2);
-        Vector4 v2 = new Vector4(2.2, 5.25, 2.0, 1.1);
+        Vector4 v1 = new Vector4(6.0f, 5.2f, 3.3f, 1.2f);
+        Vector4 v2 = new Vector4(2.2f, 5.25f, 2.0f, 1.1f);
 
-        double resultFromMethod = mathForVectors4.scalarProduct(v1, v2);
-        double result = 6.0 * 2.2 + 5.2 * 5.25 + 3.3 * 2.0 + 1.2 * 1.1;
+        float resultFromMethod = mathForVectors4.scalarProduct(v1, v2);
+        float result = 6.0f * 2.2f + 5.2f * 5.25f + 3.3f * 2.0f + 1.2f * 1.1f;
 
         double eps = 1e-6;
 
@@ -108,11 +108,11 @@ class MathForVectors4Test {
     @Test
     void normalization() {
         //Vector4 v1 = new Vector4(2.0, 4.0, 5.0, 6.0);
-        Vector4 v1 = new Vector4(1.0, 1.0, 5.0, 3.0);
+        Vector4 v1 = new Vector4(1.0f, 1.0f, 5.0f, 3.0f);
 
         Vector4 vectorFromMethod = mathForVectors4.normalization(v1);
         //Vector4 vectorResult = new Vector4(2.0 / 9.0, 4.0 / 9.0, 5.0 / 9.0, 6.0 / 9.0);
-        Vector4 vectorResult = new Vector4(1.0 / 6.0, 1.0 / 6.0, 5.0 / 6.0, 3.0 / 6.0);
+        Vector4 vectorResult = new Vector4(1.0f / 6.0f, 1.0f / 6.0f, 5.0f / 6.0f, 3.0f / 6.0f);
 
         double eps = 1e-6;
